filesystem: recognize dirctories with .lecturemetadata.dcxml as lecture root
In the past, only a directory not starting with a chapter prefix like "k" and
containing directories starting with a chapter prefix were treated as root
directories. In very rare occasions, there might be no chapters yet or the
directories start with a prefix which is unknown or no prefix at all. For this,
it is also checked whether a configuration file exists.