(G)laTeX errors: handddle edge case of unclosed math environments
If a formula was errorneous, MAGSBS tries to determine its position, by
enumerating the formulas in the document, so does GladTeX. By comparing
the numbers, it's able to emit the line number in the Markdown document.
This however doesn't work, if math environments have not been closed
properly. For this case, a new error message has been introduced.