Use stand-alone version of GladTeX
It's very difficult to get matuc and GladTeX binaries compiled with the
exact same Python version on the same environment. It can lead to hard
to trace interpreter crashes when shipping the software. Therefore,
GladTeX is now taken as a stand-alone binary. This makes the overall
binary a bit larger again, but the installer for Windows is already
fairly big with ~80M.