Add position on line for references, parsing bug fixed
This commit adds the position of reference on line - informations is
stored as an attribure in Reference class and filled during parsing.
Moreover, this information is added to the ErrorMessage in the
LinkChecker class.  This commit also fixed the bug that appears
while parsing references that contains links with titles.
Finally, this updates the __clear_link class to remove the end of lines
in the links (this behavior is similar to the pandoc generation of href
attribute)