Actually apply content filters during EPUB generation
Similarly to the previous fix for HTML (`1ca094d54`), the content filters
for the JSON AST were executed but their result not retained in the final
EPUB, as the `__apply_filters` method was designed to modify its
`json_ast` in place but didn't actually do so. This is now fixed.