Fix EPUB formula inclusion (#101)
Due to incorrect usage of `contentfilter.convert_formulas` and a wrongly
set root path, the formulas were converted to images in the parent
directory of the project root and not found by `pandoc`, so they weren't
included in the generated EPUB. These have been corrected, so the EPUB
should now also include formulas.

Note, that after this change the generated EPUB file is no longer placed
in the current working directory of the invocation but in the conversion
root (where the `.lecture_meta_data.dcxml` is located) instead.

Fixes #101.